+++ core/base/src/main/kotlin/com/rahim/yadino/base/reminder/ReminderScheduler.kt
@@ -2,7 +2,7 @@ package com.rahim.yadino.base.reminder
 
 interface ReminderScheduler {
 
-  fun setReminder(reminderName: String, reminderId: Int, reminderTime: Long, reminderIdAlarm: Long): ReminderState
+  fun setReminder(reminderName: String, reminderId: Int, reminderTime: Long, reminderIdAlarm: Int): ReminderState
 
-  suspend fun cancelReminder(reminderAlarmId: Long)
+  suspend fun cancelReminder(reminderAlarmId: Int)
 }
